Split full_node.py into 5 focused modules
The 3390-line full_node.py has been split into:
- full_node.py (~760 lines) - Core class, lifecycle, properties, connections
- full_node_block_processing.py (~1330 lines) - Block/unfinished block addition and validation
- full_node_sync.py (~825 lines) - Short/long sync, weight proof validation
- full_node_transactions.py (~250 lines) - Transaction add/broadcast
- full_node_compact_vdf.py (~445 lines) - Compact VDF operations

Method implementations are defined as module-level functions with explicit
self: FullNode parameter, then assigned as class attributes on FullNode.
PeakPostProcessingResult and WalletUpdate are re-exported from full_node.py
for backward compatibility.
